Delete GroupItem entity Add mappers A little refactor

+++ src/main/java/com/example/splitmoneybot/entity/Item.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -22,5 +21,5 @@ public class Item {
 
     private String name;
 
-    private BigDecimal money;
+    private Integer money;
 }
